+++ tests/phpunit/Unit/BaseAssetTest.php
@@ -10,13 +10,15 @@ class BaseAssetTest extends AbstractTestCase
 
     public function testBasic()
     {
+        $expectedHandle = 'expectedHandle';
+        $expectedUrl = 'expectedUrl.js';
+
         /** @var BaseAsset $testee */
-        $testee = $this->getMockForAbstractClass(BaseAsset::class);
+        $testee = $this->createTestee($expectedHandle, $expectedUrl);
 
         static::assertInstanceOf(Asset::class, $testee);
-        static::assertEmpty($testee->url());
-        static::assertEmpty($testee->handle());
-        static::assertEmpty($testee->version());
+        static::assertSame($expectedUrl, $testee->url());
+        static::assertSame($expectedHandle, $testee->handle());
         static::assertTrue($testee->enqueue());
         static::assertEmpty($testee->filters());
         static::assertEmpty($testee->data());
@@ -26,17 +28,58 @@ class BaseAssetTest extends AbstractTestCase
     public function testVersion()
     {
         /** @var BaseAsset $testee */
-        $testee = $this->getMockForAbstractClass(BaseAsset::class);
+        $testee = $this->createTestee();
+
+        // if automatic discovering of version is disabled and no version is set --> ''
+        $testee->disableAutodiscoverVersion();
         static::assertSame('', $testee->version());
 
+        $expectedFilePath = __DIR__.'/../../fixtures/style.css';
+        $expected = (string) filemtime($expectedFilePath);
+        $testee->enableAutodiscoverVersion();
+        $testee->withFilePath($expectedFilePath);
+        static::assertSame($expected, $testee->version());
+
+        // if we set a version, the version should be returned.
         $testee->withVersion('foo');
         static::assertEquals('foo', $testee->version());
+    }
+
+    /**
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     */
+    public function testFilePath()
+    {
+        $expectedFilePath = __DIR__.'/../../fixtures/style.css';
+
+        $mockedAssetPathResolver = \Mockery::mock('alias:Inpsyde\Assets\AssetPathResolver');
+        $mockedAssetPathResolver->shouldReceive('resolve')
+            ->andReturn($expectedFilePath);
+
+        $testee = $this->createTestee('foo', 'https://localhost.com/style.css');
+        static::assertSame($expectedFilePath, $testee->filePath());
+        static::assertSame($expectedFilePath, $testee->filePath());
+    }
+
+    /**
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     */
+    public function testFilePathFails()
+    {
+        $mockedAssetPathResolver = \Mockery::mock('alias:Inpsyde\Assets\AssetPathResolver');
+        $mockedAssetPathResolver->shouldReceive('resolve')
+            ->andReturn(null);
+
+        $testee = $this->createTestee('foo', 'https://localhost.com/style.css');
+        static::assertSame('', $testee->filePath());
     }
 
     public function testDependencies()
     {
         /** @var BaseAsset $testee */
-        $testee = $this->getMockForAbstractClass(BaseAsset::class);
+        $testee = $this->createTestee();
         static::assertEmpty($testee->dependencies());
 
         $testee->withDependencies('foo');
@@ -53,17 +96,17 @@ class BaseAssetTest extends AbstractTestCase
     public function testLocation()
     {
         /** @var BaseAsset $testee */
-        $testee = $this->getMockForAbstractClass(BaseAsset::class);
+        $testee = $this->createTestee();
         static::assertSame(Asset::FRONTEND, $testee->location());
 
         $testee->forLocation(Asset::BACKEND);
-        static::assertSame(ASset::BACKEND, $testee->location());
+        static::assertSame(Asset::BACKEND, $testee->location());
     }
 
     public function testFilters()
     {
         /** @var BaseAsset $testee */
-        $testee = $this->getMockForAbstractClass(BaseAsset::class);
+        $testee = $this->createTestee();
         static::assertEmpty($testee->filters());
 
         $expectedFilter1 = static function (): string {
@@ -80,7 +123,7 @@ class BaseAssetTest extends AbstractTestCase
     public function testEnqueue()
     {
         /** @var BaseAsset $testee */
-        $testee = $this->getMockForAbstractClass(BaseAsset::class);
+        $testee = $this->createTestee();
         static::assertTrue($testee->enqueue());
 
         $testee->canEnqueue(false);
@@ -99,7 +142,7 @@ class BaseAssetTest extends AbstractTestCase
         $expected = 'foo';
 
         /** @var BaseAsset $testee */
-        $testee = $this->getMockForAbstractClass(BaseAsset::class);
+        $testee = $this->createTestee();
         static::assertEmpty($testee->data());
 
         $testee->withCondition($expected);
@@ -130,5 +173,19 @@ class BaseAssetTest extends AbstractTestCase
 
         $testee->useHandler($expected);
         static::assertSame($expected, $testee->handler());
+    }
+
+    /**
+     * @param string $handle
+     * @param string $url
+     *
+     * @return \PHPUnit\Framework\MockObject\MockObject|BaseAsset
+     * @throws \InvalidArgumentException
+     * @throws \PHPUnit\Framework\Exception
+     * @throws \ReflectionException
+     */
+    private function createTestee(string $handle = 'foo', string $url = 'foo.js')
+    {
+        return $this->getMockForAbstractClass(BaseAsset::class, [$handle, $url]);
     }
 }