+++ CHANGELOG.md
@@ -2,10 +2,31 @@
 
 ## 1.4 (NOT RELEASED)
 ### New
+
+* Added `OutputFilter\InlineAssetOutputFilter` to allow printing the Asset inline.
+
+#### Automatically discovering Asset version
+Added new automatic version handling based on the `filemtime`. This can be enabled/disabled via
+* `BaseAsset::enableAutodiscoverVersion`
+* `BaseAsset::disableAutodiscoverVersion`
+
+By default, this is enabled. When no `Asset::withFilePath` or `filePath` configuration is given, the `BaseAsset` tries to solve the path to the file automatically. 
+
+Added new methods: 
+* `Asset::withFilePath` - set a filePath whichi s used in automatic discovering the version
+* `Asset::filePath` - will automatically discover the filePath if not set
+
+#### AssetManager
 * Added `AssetManager::asset(string $handle, string $type)` to access registered assets.
+
+#### PHP Version
 * Moved to PHPUnit 7.*.
 * Moved PHP min version to 7.1.
+
+#### Style/Script mutable
 * `Script` and `Style` are now mutable. See [./docs/02 - Assets.md](docs/02 - Assets.md).
+
+#### Style/Script new API features
 * `Script` now supports ...
     * inline append/prepend scripts.
     * translations for Gutenberg Blocks.
@@ -13,7 +34,6 @@
 * `Style` now supports ...   
     * append inline styles
     * set directly `Style::useAsyncFilter()` to attach filter.
-
 
 ## 1.3.1
 ### Fixes